Product API created

+++ src/main/java/com/qrzlo/webshop/data/repository/CategoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Set;
 
-public interface CategoryRepository extends CrudRepository<Category, Long>
+public interface CategoryRepository extends CrudRepository<Category, Integer>
 {
 	Set<Category> findCategoriesByNameContaining(String partialName);
 	Set<Category> findCategoriesByParent(Category parent);
